+++ dio/test/timeout_test.dart
@@ -12,7 +12,7 @@ void main() {
     dio.options.baseUrl = 'https://httpbun.com/';
   });
 
-  test('catch DioException when connect timeout', () {
+  test('catch DioException when connectTimeout', () {
     dio.options.connectTimeout = Duration(milliseconds: 3);
 
     expectLater(
@@ -26,23 +26,36 @@ void main() {
     );
   });
 
-  test('catch DioException when receiveTimeout', () {
-    dio.options.receiveTimeout = Duration(milliseconds: 10);
+  test('catch DioException when receiveTimeout', () async {
+    dio.options.receiveTimeout = Duration(seconds: 1);
 
-    expectLater(
+    final matcher = allOf([
+      throwsA(isA<DioException>()),
+      throwsA(
+        predicate<DioException>(
+          (e) => e.type == DioExceptionType.receiveTimeout,
+        ),
+      ),
+      throwsA(
+        predicate<DioException>((e) => e.message!.contains('0:00:01.000000')),
+      ),
+    ]);
+    await expectLater(
       dio.get(
-        '/bytes/${1024 * 1024 * 20}',
+        '/drip',
+        queryParameters: {'delay': 2},
+      ),
+      matcher,
+    );
+    await expectLater(
+      dio.get(
+        '/drip',
+        queryParameters: {'delay': 0, 'duration:': 1},
         options: Options(responseType: ResponseType.stream),
       ),
-      allOf([
-        throwsA(isA<DioException>()),
-        throwsA(predicate(
-            (DioException e) => e.type == DioExceptionType.receiveTimeout)),
-        throwsA(predicate(
-            (DioException e) => e.message!.contains('0:00:00.010000'))),
-      ]),
+      matcher,
     );
-  }, testOn: 'vm');
+  });
 
   test('no DioException when receiveTimeout > request duration', () async {
     dio.options.receiveTimeout = Duration(seconds: 5);
@@ -76,25 +89,51 @@ void main() {
       BaseOptions(
         baseUrl: 'https://httpbun.com/',
         connectTimeout: Duration.zero,
-        sendTimeout: Duration.zero,
         receiveTimeout: Duration.zero,
       ),
     );
     // Ignores zero duration timeouts from the base options.
     await dio.get('/drip-lines?delay=1');
     // Reset the base options.
-    dio.options
-      ..connectTimeout = Duration(seconds: 10)
-      ..sendTimeout = Duration(seconds: 10)
-      ..receiveTimeout = Duration(seconds: 10);
+    dio.options.receiveTimeout = Duration(milliseconds: 10);
+    await expectLater(
+      dio.get('/drip-lines?delay=1'),
+      allOf([
+        throwsA(isA<DioException>()),
+        throwsA(
+          predicate<DioException>(
+            (e) => e.type == DioExceptionType.receiveTimeout,
+          ),
+        ),
+        throwsA(
+          predicate<DioException>(
+            (e) => e.message!.contains('0:00:00.010000'),
+          ),
+        ),
+      ]),
+    );
+    dio.options.connectTimeout = Duration(milliseconds: 10);
+    await expectLater(
+      dio.get('/drip-lines?delay=1'),
+      allOf([
+        throwsA(isA<DioException>()),
+        throwsA(
+          predicate<DioException>(
+            (e) => e.type == DioExceptionType.connectionTimeout,
+          ),
+        ),
+        throwsA(
+          predicate<DioException>(
+            (e) => e.message!.contains('0:00:00.010000'),
+          ),
+        ),
+      ]),
+    );
+    dio.options.connectTimeout = Duration.zero;
     // Override with request options.
     await dio.get(
       '/drip-lines?delay=1',
-      options: Options(
-        sendTimeout: Duration.zero,
-        receiveTimeout: Duration.zero,
-      ),
+      options: Options(receiveTimeout: Duration.zero),
     );
-    await dio.get('/drip-lines?delay=1');
   });
 }