+++ plugins/http2_adapter/test/http2_test.dart
@@ -4,9 +4,15 @@ import 'package:dio_http2_adapter/dio_http2_adapter.dart';
 import 'package:test/test.dart';
 
 void main() {
+  test('works with non-TLS requests', () async {
+    final dio = Dio()..httpClientAdapter = Http2Adapter(ConnectionManager());
+    await dio.get('http://flutter.cn/');
+    await dio.get('https://flutter.cn/non-exist-destination');
+  });
+
   test('adds one to input values', () async {
     final dio = Dio()
-      ..options.baseUrl = 'https://pub.dev/'
+      ..options.baseUrl = 'https://httpbun.com/'
       ..interceptors.add(LogInterceptor())
       ..httpClientAdapter = Http2Adapter(
         ConnectionManager(
@@ -16,7 +22,7 @@ void main() {
       );
 
     Response<String> response;
-    response = await dio.get('?xx=6');
+    response = await dio.get('get');
     assert(response.statusCode == 200);
     response = await dio.get(
       'nkjnjknjn.html',