+++ plugins/http2_adapter/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-*None.*
+- Support non-TLS connection requests.
+- Improve the implementation of `receiveTimeout`.
 
 ## 2.3.2