+++ plugins/native_dio_adapter/CHANGELOG.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
-*None.*
+- Adds `createCronetEngine` and `createCupertinoConfiguration`
+  to deprecate `cronetEngine` and `cupertinoConfiguration`
+  for the `NativeAdapter`, to avoid platform exceptions.
+- Improve the request stream byte conversion.
 
 ## 1.1.1
 
@@ -13,6 +16,7 @@
 - Bump `cronet_http` version.
 - Minimal required Dart version is now 3.1.
 - Minimal required Flutter version is now 3.13.0.
+- Allow case-sensitive header keys with the `preserveHeaderCase` flag through options.
 
 ## 1.0.0+2